Clang Mac Build
With some help from GPT, it seems to properly build now! Some errors were just resolution-related, in regards to headers and things like that. A lot of it was compiler configuration too. I think it's because on here, I am using Clang, and on the original setup it is using msvc I think, which has certain strict parts disabled, things like that.

One of the issues was in the lceLIB library, so I had to make a commit for that to carry over. Note that you currently for this commit would have to apply this change locally yourself, until this gets upstreamed, then re-added here properly, with submodules. Had to comment out some of the tests, those need proper fixes down the road.

https://github.com/Offroaders123/lceLIB/commit/8a63b30c0ab224d14a2468cd86b40b94d9eb4ed5