Complete macOS Config
Okay now this works with compiling on macOS, as well as with VSCode IntelliSense too. It looks like you need the compile commands file for it to know how to locate things in the editor. There are still a few errors, and I don't know what Ninja is, or if I need it. That was another hangup with compiling.