Ninja Test
Found out about Ninja, it does seem to kind of work at least. There are some errors though, so I'm not sure it is quite the right option. I can't tell if that's what's being used over on the Windows version.

Note, I had to install it with Homebrew first, to use it; understandably.

Might revert this commit.

https://ninja-build.org/manual.html
https://formulae.brew.sh/formula/ninja
https://ninja-build.org/